hiver_response

Module response

Response module 响应模块

Overview / 概述

This module provides HTTP response types and utilities. 本模块提供HTTP响应类型和工具。

Structs

Response

HTTP response HTTP响应

ResponseBuilder

Response builder 响应构建器

Enums

HeaderVal

Enum for different header value types 不同头部值类型的枚举

Traits

IntoHeaderName

Trait for types that can be converted to header names 可转换为头部名称的类型的trait

IntoHeaderVal

Trait for types that can be converted to header values 可转换为头部值的类型的trait

IntoResponse

Trait for types that can be converted to HTTP responses 可转换为HTTP响应的类型的trait

Functions

status

Create a response with the given status code 使用给定状态码创建响应

Type Aliases

Body

HTTP response body HTTP响应体