hiver_response::response

Trait IntoResponse

pub trait IntoResponse {
    // Required method
    fn into_response(self) -> Response;
}

Trait for types that can be converted to HTTP responses 可转换为HTTP响应的类型的trait

Types implementing this trait can be used as return types from handlers. 实现此trait的类型可以用作处理器的返回类型。

Example / 示例

use hiver_response::IntoResponse;

struct User {
    name: String,
}

impl IntoResponse for User {
    fn into_response(self) -> Response {
        Response::builder()
            .body(format!(r#"{{"name":"{}"}}"#, self.name))
            .unwrap()
    }
}

Required Methods

fn into_response(self) -> Response

Convert self into a response 将self转换为响应